split parsing utils into separate file
