+++ cli/src/lib.rs
@@ -6,3 +6,6 @@ pub mod commands;
 
 /// Database Utils
 pub mod db;
+
+/// General cli Utils
+pub mod utils;
